Refactor: Moved HubCommunityScroll into components folder and utilized mock data
