+++ frontend/src/mock/GymHub/mockCommunities.js
@@ -1,0 +1,20 @@
+const mockCommunities = [
+    {
+        name: 'UF',
+        logo: require('../../mock/GymHub/uflogo.jpg'),
+    },
+    {
+        name: 'Workouts',
+        logo: require('../../mock/GymHub/workoutlogo.jpg'),
+    },
+    {
+        name: 'Cardio',
+        logo: require('../../mock/GymHub/cardiologo.jpg'),
+    },
+    {
+        name: 'Yoga',
+        logo: require('../../mock/GymHub/yogalogo.jpg'),
+    },
+]
+
+export default mockCommunities;
